cambios

+++ back/src/main/java/com/edu/ue/entity/UserEntity.java
@@ -2,7 +2,7 @@ package com.edu.ue.entity;
 
 public class UserEntity {
 
-    private Integer id;
+    private Long id;
     private String nombre;
     private String apellido;
     private String usuario;
@@ -18,7 +18,7 @@ public class UserEntity {
     public UserEntity() {
     }
 
-    public UserEntity(Integer id, String nombre, String apellido, String usuario, String correo, String contraseña,
+    public UserEntity(Long id, String nombre, String apellido, String usuario, String correo, String contraseña,
             String bannerImagen, String perfilImagen, String bio, String ubicacion, String sitioWebPersonal,
             String fechaNacimiento) {
         this.id = id;
@@ -35,11 +35,11 @@ public class UserEntity {
         this.fechaNacimiento = fechaNacimiento;
     }
 
-    public Integer getId() {
+    public Long  getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
